# openapi/requestor.json
{
  "components": {
    "schemas": {
      "CloneVMCommand": {
        "properties": {
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "stream_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Id"
          }
        },
        "required": [
          "name"
        ],
        "title": "CloneVMCommand",
        "type": "object"
      },
...
      "CreateVMCommand": {
        "properties": {
          "cpu": {
            "minimum": 1.0,
            "title": "Cpu",
            "type": "integer"
          },
          "memory": {
            "minimum": 1.0,
            "title": "Memory",
            "type": "integer"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "provider_id": {
            "minLength": 1,
            "title": "Provider Id",
            "type": "string"
          },
          "ssh_key": {
            "minLength": 1,
            "title": "Ssh Key",
            "type": "string"
          },
          "storage": {
            "minimum": 1.0,
            "title": "Storage",
            "type": "integer"
          },
          "stream_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Id"
          }
        },
        "required": [
          "name",
          "provider_id",
          "cpu",
          "memory",
          "storage",
          "ssh_key"
        ],
        "title": "CreateVMCommand",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResizeVMCommand": {
        "properties": {
          "cpu": {
            "minimum": 1.0,
            "title": "Cpu",
            "type": "integer"
          },
          "memory": {
            "minimum": 1.0,
            "title": "Memory",
            "type": "integer"
          },
          "storage": {
            "minimum": 1.0,
            "title": "Storage",
            "type": "integer"
          },
          "stream_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Id"
          }
        },
        "required": [
          "cpu",
          "memory",
          "storage"
        ],
        "title": "ResizeVMCommand",
        "type": "object"
      },
...
      "SnapshotCommand": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "name": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "SnapshotCommand",
        "type": "object"
      },
...
      "VMCreateResult": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider_ip": {
            "title": "Provider Ip",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "vm_id": {
            "title": "Vm Id",
            "type": "string"
          }
        },
        "required": [
          "name",
          "provider_ip",
          "vm_id",
          "config",
          "status"
        ],
        "title": "VMCreateResult",
        "type": "object"
      },
...
      "VMRecord": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider_ip": {
            "title": "Provider Ip",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "vm_id": {
            "title": "Vm Id",
            "type": "string"
          }
        },
        "required": [
          "name",
          "provider_ip",
          "vm_id",
          "config",
          "status"
        ],
        "title": "VMRecord",
        "type": "object"
      },
      "VMSnapshot": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "vm_id": {
            "title": "Vm Id",
            "type": "string"
          }
        },
        "required": [
          "name",
          "vm_id"
        ],
        "title": "VMSnapshot",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/vms": {
      "get": {
        "operationId": "list_vms_api_v1_vms_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VMRecord"
                  },
                  "title": "Response List Vms Api V1 Vms Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Vms"
      },
      "post": {
        "operationId": "create_vm_api_v1_vms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVMCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VMCreateResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Vm"
      }
    },
    "/api/v1/vms/{name}": {
      "delete": {
        "operationId": "delete_vm_api_v1_vms__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Vm"
      },
      "get": {
        "operationId": "get_vm_api_v1_vms__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VMRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Vm"
      }
    },
    "/api/v1/vms/{name}/clone": {
      "post": {
        "operationId": "clone_vm_api_v1_vms__name__clone_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloneVMCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VMRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clone Vm"
      }
    },
    "/api/v1/vms/{name}/resize": {
      "post": {
        "operationId": "resize_vm_api_v1_vms__name__resize_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResizeVMCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VMRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resize Vm"
      }
    },
    "/api/v1/vms/{name}/restart": {
      "post": {
        "operationId": "restart_vm_api_v1_vms__name__restart_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VMRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restart Vm"
      }
    },
    "/api/v1/vms/{name}/resume": {
      "post": {
        "operationId": "resume_vm_api_v1_vms__name__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VMRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Vm"
      }
    },
    "/api/v1/vms/{name}/snapshots": {
      "get": {
        "operationId": "list_snapshots_api_v1_vms__name__snapshots_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VMSnapshot"
                  },
                  "title": "Response List Snapshots Api V1 Vms  Name  Snapshots Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Snapshots"
      },
      "post": {
        "operationId": "create_snapshot_api_v1_vms__name__snapshots_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VMSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Snapshot"
      }
    },
    "/api/v1/vms/{name}/snapshots/{snapshot_name}": {
      "delete": {
        "operationId": "delete_snapshot_api_v1_vms__name__snapshots__snapshot_name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "snapshot_name",
            "required": true,
            "schema": {
              "title": "Snapshot Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Snapshot"
      }
    },
    "/api/v1/vms/{name}/snapshots/{snapshot_name}/restore": {
      "post": {
        "operationId": "restore_snapshot_api_v1_vms__name__snapshots__snapshot_name__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "snapshot_name",
            "required": true,
            "schema": {
              "title": "Snapshot Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VMRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Snapshot"
      }
    },
    "/api/v1/vms/{name}/start": {
      "post": {
        "operationId": "start_vm_api_v1_vms__name__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VMRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Vm"
      }
    },
    "/api/v1/vms/{name}/stop": {
      "post": {
        "operationId": "stop_vm_api_v1_vms__name__stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VMRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop Vm"
      }
    },
...
    "/api/v1/vms/{name}/suspend": {
      "post": {
        "operationId": "suspend_vm_api_v1_vms__name__suspend_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VMRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Suspend Vm"
      }
    },
...
  }
}